+++ tests/test_cli.py
@@ -74,3 +74,28 @@ def test_getCli_run(monkeypatch, data_dir):
     assert "python-3.9" in Results.command
     assert "--stats" in Results.command
     assert Results.env.get("LXMCUSTOM") == "1"
+
+
+def test_getCli_command(monkeypatch, data_dir):
+    import subprocess
+
+    class Results:
+        command: list[str] = None
+        env: dict[str, str] = None
+
+    def patched_subprocess(command, shell, env, *args, **kwargs):
+        Results.command = command
+        Results.env = env
+        return subprocess.CompletedProcess(command, 0)
+
+    monkeypatch.setenv(kenvmanager.KENV_PROFILE_PATH_ENV_VAR, str(data_dir))
+    monkeypatch.setattr(subprocess, "run", patched_subprocess)
+
+    extra_command = ["echo", "a", "bunch", "of", "ÅÍÎÏ˝ÓÔÒÚÆ☃", "--debug"]
+    argv = ["run", "system-test", "--"] + extra_command
+    cli = kenvmanager.get_cli(argv=argv)
+    with pytest.raises(SystemExit):
+        cli.execute()
+
+    assert Results.command == extra_command
+    assert Results.env["HEH"] == "(╯°□°）╯︵ ┻━┻)"