+++ tests/test_managers_rezenv.py
@@ -1,7 +1,17 @@
 import os
 import subprocess
 
+import pytest
+
 import kenvmanager.managers
+
+
+def test_RezEnvManager_required_fields():
+    asdict = {"params": ["--verbose"]}
+
+    with pytest.raises(ValueError) as error:
+        manager = kenvmanager.managers.RezEnvManager.from_dict(asdict)
+        assert "required field" in error
 
 
 def test_RezEnvManager_environ(monkeypatch, tmp_path):
@@ -11,8 +21,10 @@ def test_RezEnvManager_environ(monkeypatch, tmp_path):
         config={},
         environ={
             "PATH": ["$PATH", "D:\\some\\path"],
-            "NOTRESOLVED": "$PATH;D:\\some\\path",
+            "NOTRESOLVED": "foo;$$PATH;D:\\some\\path",
             "NUMBER": 1,
+            "ANOTHERONE": "$__TEST__",
+            "SUCCESSIVE": "$NUMBER",
         },
     )
 
@@ -25,6 +37,7 @@ def test_RezEnvManager_environ(monkeypatch, tmp_path):
         Results.command = command
         return subprocess.CompletedProcess(command, 0)
 
+    monkeypatch.setenv("__TEST__", "SUCCESS")
     monkeypatch.setattr(subprocess, "run", patched_subprocess)
 
     manager.execute(tmpdir=tmp_path)
@@ -33,5 +46,7 @@ def test_RezEnvManager_environ(monkeypatch, tmp_path):
     assert len(Results.env["PATH"]) > len("D:\\some\\path") + 2
     assert Results.env["PATH"].endswith(f"{os.pathsep}D:\\some\\path")
 
-    assert Results.env["NOTRESOLVED"] == "$PATH;D:\\some\\path"
+    assert Results.env["NOTRESOLVED"] == "foo;$PATH;D:\\some\\path"
     assert Results.env["NUMBER"] == "1"
+    assert Results.env["ANOTHERONE"] == "SUCCESS"
+    assert Results.env["SUCCESSIVE"] == "1"